Safari-611.1.10/LayoutTests/platform/wpe/imported/w3c/web-platform-tests/fetch/api/redirect/redirect-count-expected.txt

CONSOLE MESSAGE: Load cannot follow more than 20 redirections
CONSOLE MESSAGE: Fetch API cannot load http://localhost:8800/fetch/api/resources/redirect.py?count=1&max_age=0&redirect_status=301&max_count=21&token=085dd644-38c6-413b-83fb-9ca7766f1129&location=..%2Fresources%2Fredirect.py&count=20 due to access control checks.
CONSOLE MESSAGE: Load cannot follow more than 20 redirections
CONSOLE MESSAGE: Fetch API cannot load http://localhost:8800/fetch/api/resources/redirect.py?count=1&max_age=0&redirect_status=302&max_count=21&token=d9c1d2ec-bfee-4ff4-a740-42a165583d6d&location=..%2Fresources%2Fredirect.py&count=20 due to access control checks.
CONSOLE MESSAGE: Load cannot follow more than 20 redirections
CONSOLE MESSAGE: Fetch API cannot load http://localhost:8800/fetch/api/resources/redirect.py?count=1&max_age=0&redirect_status=303&max_count=21&token=0e3f90ff-c422-4f76-8728-d27f59a50b88&location=..%2Fresources%2Fredirect.py&count=20 due to access control checks.
CONSOLE MESSAGE: Load cannot follow more than 20 redirections
CONSOLE MESSAGE: Fetch API cannot load http://localhost:8800/fetch/api/resources/redirect.py?count=1&max_age=0&redirect_status=307&max_count=21&token=d27d93b6-3f48-49b1-836e-844131a4104d&location=..%2Fresources%2Fredirect.py&count=20 due to access control checks.
CONSOLE MESSAGE: Load cannot follow more than 20 redirections
CONSOLE MESSAGE: Fetch API cannot load http://localhost:8800/fetch/api/resources/redirect.py?count=1&max_age=0&redirect_status=308&max_count=21&token=cac43f8e-995f-405c-b1fa-7ef692e6dc39&location=..%2Fresources%2Fredirect.py&count=20 due to access control checks.

PASS Redirect 301 20 times
PASS Redirect 301 21 times
PASS Redirect 302 20 times
PASS Redirect 302 21 times
PASS Redirect 303 20 times
PASS Redirect 303 21 times
PASS Redirect 307 20 times
PASS Redirect 307 21 times
PASS Redirect 308 20 times
PASS Redirect 308 21 times
